Remove screen message from has_player_lost method
This method is a boolean and it's only responsability is to return true or false if the player has run out of guessing chances.

The screen message for win or lose logic will be at the start_game method. This way I can have all the game logic in one place.